Add svelte components
