+++ kit/src/lib/PipelineIcons/IconAudioClassification.svelte
