+++ kit/src/lib/PipelineIcons/IconAutomaticSpeechRecognition.svelte
