+++ kit/src/lib/PipelineIcons/IconConversational.svelte
