+++ kit/src/lib/PipelineIcons/IconDocumentQuestionAnswering.svelte
