+++ kit/src/lib/PipelineIcons/IconFeatureExtraction.svelte
