+++ kit/src/lib/PipelineIcons/IconFillMask.svelte
