+++ kit/src/lib/PipelineIcons/IconImageClassification.svelte
