+++ kit/src/lib/PipelineIcons/IconImageSegmentation.svelte
