+++ kit/src/lib/PipelineIcons/IconImageToImage.svelte
