+++ kit/src/lib/PipelineIcons/IconImageToText.svelte
