+++ kit/src/lib/PipelineIcons/IconObjectDetection.svelte
