+++ kit/src/lib/PipelineIcons/IconQuestionAnswering.svelte
