+++ kit/src/lib/PipelineIcons/IconRobotics.svelte
