+++ kit/src/lib/PipelineIcons/IconSentenceSimilarity.svelte
