+++ kit/src/lib/PipelineIcons/IconSummarization.svelte
