+++ kit/src/lib/PipelineIcons/IconTableQuestionAnswering.svelte
