+++ kit/src/lib/PipelineIcons/IconTabularClassification.svelte
