+++ kit/src/lib/PipelineIcons/IconTabularRegression.svelte
