+++ kit/src/lib/PipelineIcons/IconText2textGeneration.svelte
