+++ kit/src/lib/PipelineIcons/IconTextClassification.svelte
