+++ kit/src/lib/PipelineIcons/IconTextGeneration.svelte
