+++ kit/src/lib/PipelineIcons/IconTextToImage.svelte
