+++ kit/src/lib/PipelineIcons/IconTokenClassification.svelte
