+++ kit/src/lib/PipelineIcons/IconUnconditionalImageGeneration.svelte
