+++ kit/src/lib/PipelineIcons/IconVoiceActivityDetection.svelte
