+++ kit/src/lib/PipelineIcons/IconZeroShotClassification.svelte
